Add `ShieldedOutput::cmstar`
This exposes the `ExtractedCommitment` of the shielded output directly,
in addition to the byte encoding exposed by `ShieldedOutput::cmstar_bytes`.
This is useful for further generalizing scanning code that constructs
note commitment tree leaves as part of its operation, allowing the
`Domain` bound to provide the necessary type information for conversion
into those leaf types.